Fix paddle movement

+++ engine/include/entity/ball_entity.h
@@ -6,28 +6,31 @@
 #include "math/vector.h"
 #include <SDL_rect.h>
 
-namespace Breakout{
+namespace Breakout {
 
-class BallEntity: public Engine::Entity{
+class BallEntity : public Engine::Entity {
 public:
-    BallEntity(float speed = 150.f, Vector2<float> position = {}, Vector2<int> size = {16});
+  BallEntity(float speed = 150.f, Vector2<float> position = {},
+             Vector2<int> size = {16});
 
-    virtual void render(Engine::Window& target_window) override;
-    virtual void process(float delta_time) override;
+  virtual void render(Engine::Window &target_window) override;
+  virtual void process(float delta_time) override;
 
-    virtual void physics_process(float delta_time, const std::vector<std::unique_ptr<Entity>>& colliders) override;
+  virtual void physics_process(
+      float delta_time,
+      const std::vector<std::unique_ptr<Entity>> &colliders) override;
 
-    virtual void set_position(const Vector2<float>& position) override;
+  virtual void set_position(const Vector2<float> &position) override;
 
 protected:
-    void on_collision(Engine::Entity* entity) override;
+  void on_collision(Engine::Entity *entity) override;
 
 private:
-    Vector2<int> m_movement_direction;
-    float m_speed;
+  Vector2<float> m_movement_direction;
+  float m_speed;
 
-    SDL_Rect m_screen_corners;
-    Engine::Image m_ball_sprite;
+  SDL_Rect m_screen_corners;
+  Engine::Image m_ball_sprite;
 };
 
-}
+} // namespace Breakout
